+++ backend/src/main/java/com/netcracker/edu/backend/repository/DaysOfWeekRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DaysOfWeekRepository extends CrudRepository<DaysOfWeek,Long> {
+public interface DaysOfWeekRepository extends CrudRepository<DaysOfWeek,Integer> {
 }
