+++ backend/src/main/java/com/netcracker/edu/backend/service/DaysOfWeekService.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 public interface DaysOfWeekService {
     DaysOfWeek saveDaysOfWeek(DaysOfWeek day);
-    Optional<DaysOfWeek> getDaysOfWeekById(Long id);
+    Optional<DaysOfWeek> getDaysOfWeekById(int id);
     Iterable<DaysOfWeek> getAllDaysOfWeek();
-    void deleteDaysOfWeek(Long id);
+    void deleteDaysOfWeek(int id);
 }
