+++ backend/src/main/java/com/netcracker/edu/backend/service/SubjectsService.java
@@ -7,10 +7,10 @@ import java.util.Optional;
 
 public interface SubjectsService {
     Subjects saveSubjects(Subjects subject);
-    Optional<Subjects> getSubjectsById(Long id);
+    Optional<Subjects> getSubjectsById(int id);
     Iterable<Subjects> getAllSubjects();
     List<Subjects> getDistinctSubject();
     List<Subjects> getAllBySubject(String subject);
     Integer[] getIdChoosenSubject(String subject);
-    void deleteSubjects(Long id);
+    void deleteSubjects(int id);
 }
